Transfere funcoes relacionadas a 'attempts' para AttemptFuncs_DAO para desafogar ResidentFuncs_DAO
